[Docs] Rephrase repo description
This commit mentions that the process is under-documented. While in theory one should be able to use the same commands for podman as docker, the end user runs into issues with things like `--link`, the initdb.sql creation process, networking, and pods. 
Explaining that the process is under-documented makes the description look slightly less annoying, to me.